db/__init__.py
from bridge_crm.db.bootstrap import initialize_database
from bridge_crm.db.engine import get_engine
from bridge_crm.db.schema import metadata

__all__ = ["get_engine", "initialize_database", "metadata"]
