requirements.txt
Flask>=3.1,<4
Flask-WTF>=1.2,<2
SQLAlchemy>=2.0,<3
alembic>=1.14,<2
psycopg2-binary>=2.9,<3
python-dotenv>=1.0,<2
PyMySQL>=1.1,<2
gunicorn>=23.0,<24
Werkzeug>=3.1,<4
